Java_个人作品
The aim of Gue55ing Game is for a person and the computer to compete against other to correctly guess a hidden number.
A game consists of four rounds. For each round a number between 1 and 100 (inclusive) is randomly generated and the players (person and computer) take turns to guess the number. The round ends when the correct guess is given or each player has had three guesses.
If a player guesses the number correctly then they are awarded points according to how many attempts were taken to guess the number. If the round ends without either player guessing correctly then the points are awarded to each player according to how close they were to the hidden number.
At the end of the four rounds the player with the highest cumulative score wins the game.